Build: Add open source badge (#395)
Header: Add open source badge

Add an 'open source' badge that let's anyone coming to the service
easily discover the source code associated with it.
Ideally we'd redirect to a page similar to the one Image Builder has
that shows the architecture and has more docs. As a starting point, the
frontend repository may be ok.