Test: improve PW tests flakiness (#452)
* fix: add retry helper, use it for navigation

This adds a `retry` helper and uses it for retrying all current
navigations, which will be useful for mitigating flakiness during
navigation in playwright tests that caused by consoledot slowness.

* feat: add helper for waiting on task pickup

This adds a helper which is gonna indefinitely wait for a specific task
to be picked up, before continuing. This will be useful for fixing the
failures of tests that are waiting for a valid status of a repo, but can
be broken by other tasks hogging up the task queue.

* fix: move a timeout to action from test

This moves a explicit timeout to the correct spot, we have a generous
default test timeout, because as we should be setting explicit timeouts
in the smallest units (i.e. actions, expects etc.).

* fix: make closePopupsIfExist not throwing

This changes the closePopupsIfExist helper to not throw errors if
multiple handlers try to close the same pop-up, it currently throws
when a handler tries to close a pop-up has been already closed by a
prior handler.

* fix+style: include PW tests in format+lint checks

This fixes our `format` and `lint` npm commands to also check the PW
test folder. Auto format and lint fix currently present violations.

* feat: add custom eslint rule for import restriction

This adds a new custom ESLint rule that will throw linting errors when
any file inside the `_playwright-tests` folder will import any local
code that is outside the PW test folder. This will be helpful for
preventing breakages in our integration tests which pickup tests for
that folder but not the other suff in the repo.